[analyzer] Remove CompilationUnitMemberTest._updateBeginToken.
The sole purpose of this method was in parser recovery tests, to
adjust `CompilationUnit.beginToken` after parser recovery, to work
around the fact that a mismatch in begin tokens caused
`ResultComparator` to report a failure in parser recovery.

Now that `ResultComparator` doesn't examine
`CompilationUnit.beginToken` (see
https://dart-review.googlesource.com/c/sdk/+/419521), this is
unnecessary.

Removing this logic allows `CompilationUnit.beginToken` to be changed
to a final field.

Change-Id: Ic2528b5ff9dd9f1aec0fade4499b78ff165eff02
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/419680
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>