analyzer: little rename of LibraryElement2Extensions
I forgot we name an extension with "Extension", not "Extensions".

Luckily, while in the public API, this was never included in a release
of the analyzer, so it doesn't need to be shipped in a major version
release.

Change-Id: Ic66c13bac3d5b246813af6e1770a0ef6543827fd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/420421
Reviewed-by: Paul Berry <paulberry@google.com>
Auto-Submit: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>