Cider. Use CiderByteStore in tests.
Simplify MemoryByteStore.

In Cider we discard values when the last use of the key is released.

OTOH, outside of Cider, we use FileByteStore, which does not immediately discard values, but might eventually when the storage directory overflows. Using the infinite MemoryByteStore in tests is better suited for most tests.

But for Cider we need its own implementation of ByteStore that reflects the discarding behaviour.

Change-Id: Ief83eabac7f4a240feabf609b0dd0d5548bfc70f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/420640
Reviewed-by: Paul Berry <paulberry@google.com>