Fine. Fix memory leak in LibraryContext for no fine-grained dependencies.
We want to keep the cache of linked bundles with associated fine-grained dependencies and library manifests, so that we don't have to deserialize them after every API affecting change to a file in transitive dependencies. In case of the analyzer/ incremental analysis it can makes analysis up to 2x faster. OTOH, it is not clear to me how to evict entries from this cache. Maybe when the driver switches to idle, so whatever we put aside when potentially invalidating libraries, and did not use when re-loading libraries, discard these? This still theoretically could lead to negative effect if the driver is continuously kept busy with work. But better then nothing.

Without fine-grained dependencies we never tried to cache anything other than just bytes. If we need to re-load a library, we create `BundleReader` from bytes. But now, as the amount of other work becomes smaller, the overhead of reading this information from bytes becomes more visible.

So, for now I just want to fix the bug for everyone, when fine-grained dependencies are not enabled.

Change-Id: I822a89b37ce870bbdfbc474e6e140a3cf070cfdc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/420741
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>