[cfe] Dot Shorthands: Fix bug with property fileOffsets.
Surfaced from this co19 test crashing because of an assert when making a null-aware element. The `fileOffset` for static tearoffs and static gets were not being properly set and caused the crash. This CL simply sets the `fileOffset` on those expressions and existing tests should pass.

https://github.com/dart-lang/co19/blob/2c7f9a6a379cfc49b3e9019d1f616d7ec9edd766/LanguageFeatures/Static-access-shorthand/non_ambiguity_A02_t01.dart

Bug: https://github.com/dart-lang/sdk/issues/59758
Change-Id: I7b81f0ee11d8a9458154f1148e92523d1ad8be99
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/420840
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>