[analyzer] getLibraryByUri returns early again
Back in 2023 an early return was added to `getLibraryByUri`
(https://dart-review.googlesource.com/c/sdk/+/304540).

This was then migrated in January
(https://dart-review.googlesource.com/c/sdk/+/403705), which
unfortunately left the early return in a non-working state.

This CL sets libraryElement on `Reference.element2` too, which makes
the early return in `getLibraryByUri` work again and adds a test so
any future regression will hopefully be caught sooner.

On most of the getFixes requests I test in http://b/407797012 this
seems to be a 2x-3x improvement.

Change-Id: I91f2122b9ade101df78428ddb3ed5c64ad29134f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/421220
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>