[analyzer] Document immediate sub-interfaces of sealed classes in `api.txt`.
Since it is a breaking change to add a new sub-interface of a sealed
class, it is useful to see when the set of sub-interfaces of each
sealed class in the analyzer public API changes.

This change adds information to `api.txt` documenting the complete set
of sub-interfaces of each sealed class.

Based on a suggestion from Konstantin
(https://dart-review.googlesource.com/c/sdk/+/420843/comment/5a1fe050_91c066d1/).

Change-Id: I2d1673f8c1a4ade9aa880a95137491614fd8e102
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/421884
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>