[linter/analyzer_public_api] Fix crash when import is unresolved.
If the lint encounters an unresolved import when trying to compute the
set of imported public elements,
`ImportDirective.libraryImport!.importedLibrary2` will be
`null`. Previously, the lint didn't account for this possibility, so
it led to a crash. The fix is to simply ignore the unresolved library;
this is safe because an unresolved import doesn't bring any elements
into the namespace.

Fixes https://github.com/dart-lang/sdk/issues/60525.

Bug: https://github.com/dart-lang/sdk/issues/60525
Change-Id: I7feeadd60d541cffc6d872532a6006eba3cce3fe
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/422320
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>