[dart2wasm] Update closure types to be not final when dynamic modules are enabled.
Dynamic modules can introduce new closure shapes that result in new subtypes of certain closure helper structs (#Vtable, #Closure, etc.). Because we need the struct definitions to match across modules, we need to assume that there may always be subtypes of these structs and not mark any of them as final.

Change-Id: If578dc463d0f80352319d0bd44e1e4cecba28bee
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/422340
Reviewed-by: Ömer Ağacan <omersa@google.com>