Elements. Restore InstanceElement.augmented
This API was removed in 7.4.0 under the assumption that its removal would not be a breaking change, because it was marked as `@experimental`. But it turns out that some clients had been published to pub that relied on it (for example `drift_dev`).

To avoid breakages in those packages, the API has been restored (by pulling `AugmentedXyz` APIs from the published version `7.3.0`, and re-implementing on top of new `XyzElement2`). However the restored API is still slated for removal in analyzer 8.0.0, so it's been marked as `@deprecated`. Clients relying on this API should migrate away from it as soon as possible.

Bug: https://github.com/dart-lang/sdk/issues/60545
Bug: https://github.com/dart-lang/sdk/issues/60553

Change-Id: If12c1fa37cafa63d0eef121a39350ce6d41ab68f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/422846
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>