[analyzer] Dot Shorthands: Add DotShorthandConstructorInvocation AST and resolve constructors.
When we resolve a DotShorthandInvocation, it can either be a method invocation or a constructor invocation. This CL adds a `DotShorthandConstructorInvocation` node to represent the shorthand equivalent of an `InstanceCreationExpression` (without the extra bulk of needing a synthetic `NamedType`), and adds the resolution logic.

Test added in the form of unit test and existing language tests.

Bug: https://github.com/dart-lang/sdk/issues/59835
Change-Id: I6966ce549f2b079efc0b002fba41bc31a90871c1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/422544
Reviewed-by: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>