analyzer: Remove unnecessary AnalysisTarget
This just seems like an odd parent class for both Element and
ConstantEvaluationTarget. I don't think it carries it's weight (added
complexity).

The abstract AnalysisTarget serves as an interface that adds only two
getters to the contract: `Source? get librarySource` and
`Source? get source`, and is only referenced by two classes: `Element`
implements it, and `ConstantEvaluationTarget` extends it (but it's
abstract anyway).

So to me it seems this is a false type hierarchy; just because Element
and ConstantEvaluationTarget both have a source and a librarySource
field does not mean that it is useful or meaningful to encode this
information in the type hierarchy.

Change-Id: I08ef8e32d3ead6a969950ff34f76baa84272b662
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/399321
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>