[analysis_server] Pre-emptively opt some tests out of newline normalization
Currently newline normalization is opt-in for server tests because there are many tests that fail when the newlines are normalized to the current platform. This change pre-emptively opts-out a number of tests that fail so that the default can be flipped, and then these tests could be worked through.

See https://github.com/dart-lang/sdk/issues/60234

Change-Id: I19f5121612fbc9da18ad0edbe631549e04553e6d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/423800
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>